requirements.txt
asgiref==3.8.1
Brotli==1.1.0
dj-database-url==2.1.0
Django==5.1.7
django-widget-tweaks==1.5.0
gunicorn==21.2.0
packaging==25.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.7.0
dj-database-url>=2.1.0

# Dependências adicionais
urllib3>=1.26.0,<3.0.0
certifi>=2024.2.2
